Move over another set of pytests and fixed.

+++ python/geocal/sqlite_shelf.py
@@ -8,21 +8,21 @@
 #
 # This class gives a Shelve like interface to sqlite.
 
-# UserDict goes away in python 3 (but is largely the same as collections). Try
-# the python 2 import first, then fall back to python 3
-try:
-    from UserDict import DictMixin
-except ImportError:
-    from collections.abc import MutableMapping as DictMixin
+from __future__ import annotations
+from collections.abc import MutableMapping as DictMixin
 
 import pickle
 import sqlite3
 import os.path
-import geocal_swig
+import geocal_swig  # type: ignore
+import os
+from pathlib import Path
+from typing import Any
+from collections.abc import Iterator
 
 # Allow jsonpickle to be used, if available.
 try:
-    import jsonpickle
+    import jsonpickle  # type: ignore
 
     have_jsonpickle = True
     jsonpickle.set_encoder_options(
@@ -32,7 +32,7 @@ except ImportError:
     have_jsonpickle = False
 
 
-def to_db_type(value):
+def to_db_type(value: Any) -> Any:
     """If value's type is supported natively in SQLite, return value.
     Otherwise, return a pickled representation."""
     if value is None or isinstance(value, (int, int, float, bytes, str)):
@@ -41,14 +41,14 @@ def to_db_type(value):
         return bytes(pickle.dumps(value))
 
 
-def from_db_type(value):
+def from_db_type(value: Any) -> Any:
     """Converts a value from the database to a Python object."""
     if isinstance(value, bytes):
         return pickle.loads(value)
     return value
 
 
-def read_shelve(f):
+def read_shelve(fn: str | os.PathLike[str]) -> Any:
     """This handles reading a value from a shelve file. The string f should
     be of the form file_name:key, file_name.xml, file_name.bin,
     or file_name.json. We
@@ -79,6 +79,7 @@ def read_shelve(f):
     one). We change back to the current directory when done.
 
     """
+    f = str(Path(fn))
     fname = f.split(":")[0]
     dirn, fb = os.path.split(fname)
     curdir = os.getcwd()
@@ -111,11 +112,13 @@ def read_shelve(f):
         os.chdir(curdir)
 
 
-def shelve_time_after(f1, f2):
+def shelve_time_after(f1n: str | os.PathLike[str], f2n: str | os.PathLike[str]) -> bool:
     """Compare the update time on 2 shelve objects, return if f1 update time
     >= f2 update time. Note that either f1 or f2 can be files, in which case
     we use the file modify time instead.
     It is ok if f1 doesn't exist, in that case always return False."""
+    f1 = str(Path(f1n))
+    f2 = str(Path(f2n))
     if ":" in f1:
         fname, key = f1.split(":")
         t = SQLiteShelf(fname, "r")
@@ -137,7 +140,7 @@ def shelve_time_after(f1, f2):
     return f1time >= f2time
 
 
-def write_shelve(f, val):
+def write_shelve(fn: str | os.PathLike[str], val: Any) -> None:
     """This handles writing a value to a shelve file, possibly creating the
     file is it doesn't exist. The string f should be of the form
     file_name:key. We open/create the given file and write the value for
@@ -148,6 +151,7 @@ def write_shelve(f, val):
     So we also support xml and bin files, we key off of the file name and if it
     is something like "foo.xml" we write that rather than a shelve file.
     """
+    f = str(Path(fn))
     if os.path.splitext(f)[1] == ".xml":
         geocal_swig.serialize_write(f, val)
         return
@@ -172,11 +176,11 @@ def write_shelve(f, val):
 class SQLiteShelf(DictMixin):
     """Shelf implementation using an SQLite3 database."""
 
-    def __init__(self, filename, mode="r+"):
+    def __init__(self, filename: str | os.PathLike[str], mode: str = "r+") -> None:
         """Open an existing file, or create a new one if it doesn't exist.
         The mode can be "r+" for read/write or "r" for read only.
         """
-        self._database = sqlite3.connect(filename)
+        self._database: None | sqlite3.Connection = sqlite3.connect(filename)
         self._database.execute(
             "CREATE TABLE IF NOT EXISTS Shelf "
             "(Key TEXT PRIMARY KEY NOT NULL, Value BLOB, Updated datetime)"
@@ -184,10 +188,12 @@ class SQLiteShelf(DictMixin):
         self._open = True
         self._read_only = mode == "r"
 
-    def __del__(self):
+    def __del__(self) -> None:
         self.close()
 
-    def __getitem__(self, key):
+    def __getitem__(self, key: str) -> Any:
+        if self._database is None:
+            raise RuntimeError("Attempt to use database after it has been closed()")
         row = self._database.execute(
             "SELECT Value FROM Shelf WHERE Key=?", [key]
         ).fetchone()
@@ -196,8 +202,10 @@ class SQLiteShelf(DictMixin):
         else:
             raise KeyError(key)
 
-    def update_time(self, key):
+    def update_time(self, key: str) -> str:
         """Return updated time as a string."""
+        if self._database is None:
+            raise RuntimeError("Attempt to use database after it has been closed()")
         row = self._database.execute(
             "SELECT Updated FROM Shelf WHERE Key=?", [key]
         ).fetchone()
@@ -206,8 +214,10 @@ class SQLiteShelf(DictMixin):
         else:
             raise KeyError(key)
 
-    def update_time_julian(self, key):
+    def update_time_julian(self, key: str) -> float:
         """Return updated time as Julian day, including fraction"""
+        if self._database is None:
+            raise RuntimeError("Attempt to use database after it has been closed()")
         row = self._database.execute(
             "SELECT julianday(Updated) FROM Shelf WHERE Key=?", [key]
         ).fetchone()
@@ -216,13 +226,15 @@ class SQLiteShelf(DictMixin):
         else:
             raise KeyError(key)
 
-    def update_time_unix(self, key):
+    def update_time_unix(self, key: str) -> float:
         """Return updated time as unix time, including fraction"""
         # Unix epoch in Julian days is 2440587.5
         return (self.update_time_julian(key) - 2440587.5) * 86400.0
 
-    def touch(self, key):
+    def touch(self, key: str) -> None:
         """Change the updated time to now (like the unix command "touch")"""
+        if self._database is None:
+            raise RuntimeError("Attempt to use database after it has been closed()")
         if self._read_only:
             raise RuntimeError("Attempt to write to read only shelve.")
         self._database.execute(
@@ -230,7 +242,9 @@ class SQLiteShelf(DictMixin):
             [key],
         )
 
-    def __setitem__(self, key, value):
+    def __setitem__(self, key: str, value: Any) -> None:
+        if self._database is None:
+            raise RuntimeError("Attempt to use database after it has been closed()")
         if self._read_only:
             raise RuntimeError("Attempt to write to read only shelve.")
         self._database.execute(
@@ -238,16 +252,20 @@ class SQLiteShelf(DictMixin):
             [key, to_db_type(value)],
         )
 
-    def __delitem__(self, key):
+    def __delitem__(self, key: str) -> None:
+        if self._database is None:
+            raise RuntimeError("Attempt to use database after it has been closed()")
         if self._read_only:
             raise RuntimeError("Attempt to delete from read only shelve.")
         self._database.execute("DELETE FROM Shelf WHERE Key=?", [key])
 
-    def keys(self):
+    def keys(self) -> list[str]:  # type: ignore[override]
         """Return a list of keys in the shelf."""
+        if self._database is None:
+            raise RuntimeError("Attempt to use database after it has been closed()")
         return [row[0] for row in self._database.execute("SELECT Key FROM Shelf")]
 
-    def close(self):
+    def close(self) -> None:
         """Commit changes and close the file."""
         if self._database is not None:
             self._database.commit()
@@ -255,10 +273,10 @@ class SQLiteShelf(DictMixin):
             self._database = None
 
     # These are needed by python 3, but not python 2
-    def __len__(self):
+    def __len__(self) -> int:
         raise RuntimeError("Not implemented yet")
 
-    def __iter__(self):
+    def __iter__(self) -> Iterator[Any]:
         raise RuntimeError("Not implemented yet")
